feat: Extract house keeper into separate crate
