+++ core/lib/zksync_core/Cargo.toml
@@ -45,6 +45,7 @@ vlog.workspace = true
 zksync_eth_watch.workspace = true
 zksync_shared_metrics.workspace = true
 zksync_commitment_generator.workspace = true
+zksync_house_keeper.workspace = true
 
 multivm.workspace = true
