+++ core/node/house_keeper/README.md
@@ -1,0 +1,11 @@
+# zkSync Era housekeeper
+
+This crate contains functionality for performing “administrative” work to keep the system flowing. It does:
+
+- reporting
+- archiving
+- retrying
+- monitoring
+- job scheduling
+- queue cleaning
+- etc.
