+++ core/node/house_keeper/src/blocks_state_reporter.rs
@@ -3,7 +3,7 @@ use zksync_dal::{ConnectionPool, Core, CoreDal};
 use zksync_shared_metrics::{BlockL1Stage, BlockStage, L1StageLatencyLabel, APP_METRICS};
 use zksync_utils::time::seconds_since_epoch;
 
-use crate::house_keeper::periodic_job::PeriodicJob;
+use crate::periodic_job::PeriodicJob;
 
 #[derive(Debug)]
 pub struct L1BatchMetricsReporter {
