+++ core/node/house_keeper/src/fri_proof_compressor_job_retry_manager.rs
@@ -4,7 +4,7 @@ use async_trait::async_trait;
 use prover_dal::{Prover, ProverDal};
 use zksync_dal::ConnectionPool;
 
-use crate::house_keeper::periodic_job::PeriodicJob;
+use crate::periodic_job::PeriodicJob;
 
 #[derive(Debug)]
 pub struct FriProofCompressorJobRetryManager {
