+++ core/node/house_keeper/src/fri_scheduler_circuit_queuer.rs
@@ -2,7 +2,7 @@ use async_trait::async_trait;
 use prover_dal::{Prover, ProverDal};
 use zksync_dal::ConnectionPool;
 
-use crate::house_keeper::periodic_job::PeriodicJob;
+use crate::periodic_job::PeriodicJob;
 
 #[derive(Debug)]
 pub struct SchedulerCircuitQueuer {
