+++ core/node/node_framework/Cargo.toml
@@ -30,6 +30,7 @@ zksync_circuit_breaker.workspace = true
 zksync_concurrency.workspace = true
 zksync_eth_watch.workspace = true
 zksync_commitment_generator.workspace = true
+zksync_house_keeper.workspace = true
 
 tracing.workspace = true
 thiserror.workspace = true
